Add support for responses to takeout doubles
This isn't perfect.  We fail a bunch of tests still
because SuitedToPlay can't respond to a TakeoutDouble
as it currently only promises 11+hcp.  It's probable that should be 12+
or suited to play should be made 11+ or shoudl check !preempt which is
I think what it's trying to avoid applying over.